weather_agent/sub_agents/location_agent.py
from google.adk import Agent
from ..tools import search_location

location_agent = Agent(
    model='gemini-2.0-flash-001',
    name='location_finder',
    description="Finds coordinates for cities.",
    instruction="""You are a location finder. Your job is to find the coordinates of a city.
    
    When you receive a request:
    1. Identify the city name from the user's prompt.
    2. Identify any context like country, region, state, or province.
    3. Use the search_location tool to find the city. 
    4. If multiple results are returned, use the context to filter and select the correct one.
    5. Return the result in a clear format.

    IMPORTANT: 
    - Don't give country, region, state, or province to the search_location tool, it only finds cities.
    - When the user gives several names like "Goya, Corrientes, Argentina" or "New York, New York, USA", the first name is usually the city name.
    - If the tool gives you multiple results, use the context to filter and select the correct ones.
    """,
    tools=[search_location]
)
